refactor(test): group remaining runtime coverage
