+++ ci/cli-test-timing-hints.json
@@ -191,7 +191,7 @@
     "test/reboot-identity-drift.test.ts": 13632,
     "test/credentials/rebuild-credential-hydration.test.ts": 9487,
     "test/credentials/rebuild-credential-preflight.test.ts": 22966,
-    "test/recover-port-forward.test.ts": 12821,
+    "test/runtime/gateway/recover-port-forward.test.ts": 12821,
     "test/repository/source-architecture.test.ts": 5535,
     "test/repository/source-require-loader.test.ts": 5155,
     "test/repro-2666-silent-list-status.test.ts": 12380,
